requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
langchain
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
python-multipart
pinecone
pinecone-client
openai